+++ src/main/java/org/programmers/signalbuddyfinal/domain/auth/controller/AuthController.java
@@ -67,7 +67,7 @@ public class AuthController {
     @PostMapping("/verify-code")
     public ResponseEntity<ApiResponse<Object>> verifyCode(
         @Valid @RequestBody VerifyCodeRequest verifyCodeRequest) {
-        emailService.verifyCode(verifyCodeRequest);
+        authService.verifyCode(verifyCodeRequest);
         return ResponseEntity.ok().body(ApiResponse.createSuccessWithNoData());
     }
